requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.0
Flask==2.3.3
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
mysql-connector-python==8.3.0